feat: require confirm for shroud an reveal all
